requirements.txt
pandas~=2.2.3
flask~=3.1.0
Werkzeug~=3.1.3
prometheus-client~=0.21.1
pm4py~=2.7.12.4
process-mining-core~=0.1.3
kafka-python-ng~=2.2.3